+++ packages/components/CHANGELOG.md
@@ -9,14 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **mi-floor-selector** is a new component for showing the "floor-selector".
-- **mi-my-position** is a new component for showing the "my-position" button.
+- **mi-floor-selector** is a new component for showing a MapsIndoors  Floor Selector.
+- - **mi-my-position** is a new component for showing a MapsIndoors My Position button.
 
 ## [13.6.0] - 2023-07-06
 
 ### Added
 
-- **mi-list-item-location** implemets `showExternalId` property.
+- **mi-list-item-location** implements `showExternalId` property.
 
 ## [13.5.0] - 2023-07-03